fix(virtio_gpu): cast to size_t before width*height to avoid 32-bit overflow
The framebuffer test-pattern loop bounds were computed as uint32_t * uint32_t
and then promoted to size_t. For displays where the product exceeds 2^32
(4096x4096x4 = 64MiB still fits, but 8K screens overflow), the multiplication
wraps before the size_t conversion. Mirrors the (size_t)width * height cast
already used at line 519 when sizing the framebuffer. Closes one
cpp/integer-multiplication-cast-to-long code-scanning alert.

+++ kernel/video/virtio_gpu.c
@@ -893,7 +893,7 @@ int virtio_gpu_init(uint64_t *out_fb_phys, uint32_t width, uint32_t height) {
     volatile uint32_t *fb = (volatile uint32_t *)g_framebuffer_guest;
     uint32_t pixel = 0xFFFFFFFF;  /* White - all channels max */
 
-    for (size_t i = 0; i < width * height; i++) {
+    for (size_t i = 0; i < (size_t)width * height; i++) {
         fb[i] = pixel;
     }
